refactor: add computer repository
