+++ Program.cs
@@ -1,33 +1,25 @@
 ﻿using Microsoft.Data.Sqlite;
 using LabManager.Database;
+using LabManager.Repositories;
 
 var databaseConfig = new DatabaseConfig();
 new DatabaseSetup(databaseConfig);
 
-// Routing
+var computerRepository = new ComputerRepository(databaseConfig);
 
+// Routing
 var modelName = args[0];
 var modelAction = args[1];
 
 if (modelName == "Computer")
 {
     if (modelAction == "List")
-    {
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();
-        var command = connection.CreateCommand();
-        command.CommandText = "SELECT * FROM Computers;";
-
-        var reader = command.ExecuteReader();
-
+    {  
         Console.WriteLine("Computer List");
-        while(reader.Read())
+        foreach (var computer in computerRepository.GetAll())
         {
-            Console.WriteLine("{0}, {1}, {2}", reader.GetInt32(0), reader.GetString(1), reader.GetString(2));
+            Console.WriteLine("{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor);
         }
-        
-        reader.Close();
-        connection.Close();
     }
 
     if (modelAction == "New")
